requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-aiplatform==1.26.0
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
gunicorn
